pdf-translator/requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyMuPDF==1.27.2
pdf2docx==0.5.12
python-docx==1.2.0
fpdf2==2.8.1
deep-translator==1.11.4
google-generativeai==0.5.4
anthropic==0.23.1
python-dotenv==1.0.0
spacy==3.7.4
requests>=2.28.0
